Push fetching reviews down to the viewmodel
